application/requirements.txt
pandas
numpy
pillow
statsmodels
tensorflow
matplotlib
openpyxl
os